server/requirements.txt
aiosqlite==0.17.0
appdirs==1.4.4
black==21.5b1
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
databases==0.4.3
decorator==4.4.2
dnspython==1.16.0
docopt==0.6.2
docutils==0.17.1
fastapi==0.65.1
fett==0.3.2
flake8==3.9.2
h11==0.12.0
idna==2.10
mccabe==0.6.1
mypy-extensions==0.4.3
networkx==2.5.1
pathspec==0.8.1
Pillow==8.2.0
pycodestyle==2.7.0
pydantic==1.8.2
pyflakes==2.3.1
pymongo==3.11.4
python-dotenv==0.17.1
python-jsonrpc-server==0.3.4
PyYAML==5.4.1
regex==2021.4.4
requests==2.25.1
snooty-lextudio==1.9.2.dev0
SQLAlchemy==1.3.24
starlette==0.14.2
toml==0.10.2
typing-extensions==3.7.4.3
ujson==1.35
urllib3==1.26.5
uvicorn==0.13.4
watchdog==1.0.2
